# src/internal/model/dbmodel/user.go
package dbmodel

import "time"

type User struct {
	ID        uint      `json:"id"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
	Name      string    `json:"name"`
	Email     string    `json:"email" gorm:"unique"`
	Password  string    `json:"password"`
} //@name User

func (User) TableName() string {
	return "user"
}